# [Backport] Fix condition Safe-NULL condition on Update Trigger