@sebastien-kathmandu 2019-05-31
I forgot to specify that it impact also the branch `2-3-develop`