@orlangur 2019-06-01
@akaplya would be _really_ nice if you can review this MView/Staging logic change.