@sebastien-kathmandu 2019-06-04
Hi @rodrigowebjump ,

We are still using 2.2-develop at work so I did it from `2.3-develop`. Do you need a rebase or can you work with it?

Cheers,