@QuentinFarizonAfrimarket 2019-06-04
Thanks @sebastien-kathmandu for this fix !

I think that triggers should be reinstalled after applying this as a patch, or upgrading to a version that contains it, no ?

In the absence of a command to do that, or even proper documentation (see https://github.com/magento/magento2/issues/23023 and https://github.com/magento/devdocs/issues/4647 ), should this PR contain a setupUpgrade script that reinstalls all triggers ? Is that even doable ?  

Anyway, a `bin/magento indexer:triggers:reinstall` is badly need I think