@sebastien-kathmandu 2019-06-04
Hi @QuentinFarizonAfrimarket ,

It's a really good point. We did in our environment by switching the indexes to `On save` and back to `On schedule` and so the triggers are regenerated. I should do a `SetupUgrade` indeed running the command that regenerate them. I'll look at it ASAP. in the same time the command line can be a good idea.

Cheers,