@paales 2019-06-07
There is `Magento\Indexer\Setup\RecurringData`, doesn't that already remove all and add all triggers?