@QuentinFarizonAfrimarket 2019-06-07
Yes, `Magento\Indexer\Setup\RecurringData` seems to do what is needed here.
Not sure if we can call it from a schema upgrade and/or a cli command.