@sebastien-kathmandu 2019-06-08
Hi @paales  and @QuentinFarizonAfrimarket ,

I looked and indeed it should do it. But when we deployed the fix into our staging and production environment, the cli `setup:upgrade` didn't update the triggers, I had to switch to `on Save` and back to `Schedule` my indexers. I'll try it on my local to see why. but hopefully you're right guys ;-).

Cheers,