@hostep 2019-06-17
@sebastien-kathmandu: it might be better to first submit a PR which targets the `2.3-develop` branch, otherwise this PR is not going to be processed very fast I'm afraid. Thanks! 🙂

(for the people interested: there is also some discussion going on around this PR and the history of this bug in https://magentocommeng.slack.com/archives/C4YS78WE6/p1560763042380000)