@mattheo-geoffray 2019-06-18
For people looking at this PR there is also this logic in a Catalog Staging class in the Commerce version