@sebastien-kathmandu 2019-06-18
Hi @mattheo-geoffray ,

Indeed in the file `src/vendor/magento/module-catalog-staging/Model/Mview/View/Attribute/Subscription.php` when using composer Line 134 in my case.

Thanks @mattheo-geoffray .

Cheers,