@VladimirZaets 2019-07-10
Hi @sebastian-virtua. The branch `2.2-develop` will be closed for contribution from **July 15, 2019**. Please finish all your Pull Requests till this date, otherwise, they will be closed.
Thanks for collaboration.