@VladimirZaets 2019-07-15
Hi @sebastien-kathmandu  
The next 2.2 release will be the last of functional releases.
Due to it, we are closing 2.2-develop for contribution and all not ready for delivery Pull Requests. 
Thanks for collaboration.